general refactor

+++ src/check_cheapest.c
@@ -12,7 +12,7 @@
 
 #include "push_swap.h"
 
-int	find_index_b(t_stacks *stacks, long num)
+int	find_index_b(t_stacks *stacks, int num)
 {
 	int				i;
 	int				size;
@@ -31,7 +31,7 @@ int	find_index_b(t_stacks *stacks, long num)
 	return (i);
 }
 
-int	find_index_a(t_stacks *stacks, long num)
+int	find_index_a(t_stacks *stacks, int num)
 {
 	int				i;
 	int				size;
@@ -50,7 +50,7 @@ int	find_index_a(t_stacks *stacks, long num)
 	return (i);
 }
 
-long	search_num_b(t_stacks *stacks, long num)
+long	search_num_b(t_stacks *stacks, int num)
 {
 	t_stack_item	*head_b;
 	int				size;
@@ -76,7 +76,7 @@ long	search_num_b(t_stacks *stacks, long num)
 	return (num);
 }
 
-long	search_num_a(t_stacks *stacks, long num)
+long	search_num_a(t_stacks *stacks, int num)
 {
 	t_stack_item	*head_a;
 	int				size;
@@ -87,13 +87,14 @@ long	search_num_a(t_stacks *stacks, long num)
 	flag = 0;
 	head_a = stacks->a_head;
 	size = ft_listsize(stacks->a_head);
-	while (flag == 0) {
+	while (flag == 0)
+	{
 		i = 0;
 		num++;
 		head_a = stacks->a_head;
 		while (i++ < size)
 		{
-			if(head_a->content == num)
+			if (head_a->content == num)
 				flag = 1;
 			head_a = head_a->next;
 		}
